Simplify class Node.java as a pure data structure, and hide it
