+++ concurrent-linked-queue-simple/src/main/java/com/bobocode/ConcurrentLinkedQueue.java
@@ -9,6 +9,19 @@ import java.util.concurrent.atomic.AtomicInteger;
  * @param <T> a generic parameter
  */
 public class ConcurrentLinkedQueue<T> implements Queue<T> {
+    static final class Node<T> {
+        private T element;
+        private Node<T> next;
+
+        static <T> Node<T> valueOf(T element) {
+            return new Node<>(element);
+        }
+
+        private Node(T element) {
+            this.element = element;
+        }
+    }
+
     private Node<T> head;
     private Node<T> tail;
     private AtomicInteger size = new AtomicInteger();
@@ -19,7 +32,7 @@ public class ConcurrentLinkedQueue<T> implements Queue<T> {
         if (head == null) {
             head = tail = newNode;
         } else {
-            tail.setNext(newNode);
+            tail.next = newNode;
             tail = newNode;
         }
         size.incrementAndGet();
@@ -28,8 +41,8 @@ public class ConcurrentLinkedQueue<T> implements Queue<T> {
     @Override
     synchronized public T poll() {
         if (head != null) {
-            T element = head.getElement();
-            head = head.getNext();
+            T element = head.element;
+            head = head.next;
             if (head == null) {
                 tail = null;
             }
